cast sizeof to int
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@774 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/test/typestest.c
@@ -38,7 +38,7 @@ main(int argc, char *argv[]) {
     printf("nrrdTypeMax: % 31.15f\n", nrrdTypeMax[tt]);
     printf("nrrdTypeNumberValues: % 31.15f\n", nrrdTypeNumberValues[tt]);
   }
-  printf("sizeof(size_t) = %d\n", sizeof(size_t));
+  printf("sizeof(size_t) = %d\n", (int)sizeof(size_t));
 
   /*
   c = -10;
